chore: Rename typescript folder
