+++ src/context/AuthContext.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import React, { createContext, useEffect, useState } from 'react';
 import { whoamiRequest } from '../services/session.services';
-import { TUser } from '../typescript/types';
+import { TUser } from '../types/types';
 
 // Create the context and set the initial values
 export const AuthContext = createContext({
